Fixed serious bug in parser generator CC code
Although itemsets in CC are each "closed", they are NOT necessarily disjoint.
The closure of one set can strictly contain a smaller set which is still
closed, so we *cannot* check if itemsets we generate in the CC algorithm are
new by just checking 1 element. We actually need to compare the sets for
equality.