Rewriting AST type
Using type system to enforce AST children / parents is not worth the enormous
number of extra types we need to create.

So just have an overall AST type and use it for all nodes, with lots of
variants for different types of AST nodes. This means that as far as the type
system is concerned anything in the AST can be a child of anything else.